Pixel buffer revamp (#22)
Changing the way the pixels are processed in the buffer.

Instead of directly manipulating the raw rows and columns, there is now a full matrix of ROWSxCOLS pixels which will get manipulated. If a point in the pixel map corresponds to a physical pixel in the display, the color will get "written through" to the physical pixel.

NOTE: The CenterOut pattern has not yet been updated to use the new buffer and still relies on the raw rows and columns. (I haven't had the time to modify the "shiftX" method overloads that take a "startingBlock" parameter.)

* Make PixelBuffer constructor protected

* Remove artificial wait after showing pixels

* Remove unused var

* Implement matrix lookup logic for shift-right

* Force 40x40 matrix for testing

* Use new pixel map for Up/Left

* Cleanup and comments

* Remove extra test rows/cols - messing up CenterOut