First refactoring to support file-based configuration (#27)
Refactoring logic in advance of adding support for file-based configuration.

Adding PixelMap as an alternative to the existing PixelBuffer. PixelMap has no dependency on the Adafruit_NeoPixel library and enables cleaner separation of dependencies.

Modified DisplayPattern base class and all child classes to accept a PixelMap in addition to the PixelBuffer. This will allow for back-compat until all sign driver logic is changed. (Currently only the LegacyController uses the new PixelMap.)

Adding DisplayConfiguration class to be able to parse JSON to determine the sign's physical layout. A hard-coded DisplayConfigFactory class currently utilizes built-in JSON to create a config based on sign type. Eventually this will be replaced by data read from the SD card.