Enable reading from SD card for sign config (#36)
Read configurations from SD card, if present.
Primary and Legacy controllers have different pinouts, so check 2 different configurations to see if an SD card reader is present.
If no card reader is present, fall back to the pin-based configurations to try to determine the sign type.
The logic works with Secondary controllers in a "stand alone" mode.