Merging Secondary logic into Legacy Controller (#37)
There are currently 3 workspaces with 3 different main.cpp files: One for the Primary controller, one for the Secondary controller, and one for the Legacy sign. The workspaces were separate because they needed slightly different logic and had different pinouts. With the recent change to allow the Legacy controller to read its configuration from a microSD card, it's now possible to merge the logic for the Secondary controller and the Legacy controller into the same workspace.

A subsequent update is planned to make the Primary controller get its configuration from a microSD card, at which point it might be possible to merge it into the common logic and have a single workspace.